# Further speed up Bundler by not sorting specs unnecessarily