## What was the end-user or developer problem that led to this PR?

Trying to make Bundler faster and simplify code at the same time.

## What is your fix for the problem, implemented in this PR?

Improve `bundler/setup` time by ~2% in the big Gemfile I commonly use for testing performance (https://github.com/technicalpickles/big-gemfile), although it does not improve simpler setups like a brand new Rails app.

```
✗ hyperfine 'BUNDLER_VERSION=2.3.22 ruby -rbundler/setup -e1' 'BUNDLER_VERSION=2.4.0.dev ruby -rbundler/setup -e1'  --warmup 3
Benchmark 1: BUNDLER_VERSION=2.3.22 ruby -rbundler/setup -e1
  Time (mean ± σ):     141.4 ms ±   1.0 ms    [User: 112.5 ms, System: 27.6 ms]
  Range (min … max):   139.9 ms … 143.1 ms    20 runs
 
Benchmark 2: BUNDLER_VERSION=2.4.0.dev ruby -rbundler/setup -e1
  Time (mean ± σ):     139.0 ms ±   2.4 ms    [User: 109.4 ms, System: 27.6 ms]
  Range (min … max):   136.8 ms … 147.6 ms    21 runs
 
Summary
  'BUNDLER_VERSION=2.4.0.dev ruby -rbundler/setup -e1' ran
    1.02 ± 0.02 times faster than 'BUNDLER_VERSION=2.3.22 ruby -rbundler/setup -e1'
```

The main idea is to not sort specs unnecessarily when not needed although there are other misc improvements in the PR.

## Make sure the following tasks are checked

- [x] Describe the problem / feature
- [ ] Write [tests](https://github.com/rubygems/rubygems/blob/master/bundler/doc/development/PULL_REQUESTS.md#tests) for features and bug fixes
- [x] Write code to solve the problem
- [x] Make sure you follow the [current code style](https://github.com/rubygems/rubygems/blob/master/bundler/doc/development/PULL_REQUESTS.md#code-formatting) and [write meaningful commit messages without tags](https://github.com/rubygems/rubygems/blob/master/bundler/doc/development/PULL_REQUESTS.md#commit-messages)